+++ backend/Controllers/DownloadController.php
@@ -66,12 +66,14 @@ class DownloadController
             // @codeCoverageIgnoreEnd
         });
 
-        $mimes = (new MimeTypes())->getMimeTypes(pathinfo($file['filename'], PATHINFO_EXTENSION));
+        $extension = pathinfo($file['filename'], PATHINFO_EXTENSION);
+        $mimes = (new MimeTypes())->getMimeTypes($extension);
         $contentType = !empty($mimes) ? $mimes[0] : 'application/octet-stream';
 
         $disposition = HeaderUtils::DISPOSITION_ATTACHMENT;
 
-        if ($contentType == 'application/pdf') {
+        $download_inline = (array)$this->config->get('download_inline', ['pdf']);
+        if (in_array($extension, $download_inline) || in_array('*', $download_inline)) {
             $disposition = HeaderUtils::DISPOSITION_INLINE;
         }
 
@@ -89,7 +91,12 @@ class DownloadController
             'Content-Transfer-Encoding',
             'binary'
         );
-
+        if (isset($file['filesize'])) {
+            $streamedResponse->headers->set(
+                'Content-Length',
+                $file['filesize']
+            );
+        }
         // @codeCoverageIgnoreStart
         if (APP_ENV == 'development') {
             $streamedResponse->headers->set(
@@ -135,11 +142,11 @@ class DownloadController
     public function batchDownloadStart(Request $request, StreamedResponse $streamedResponse, TmpfsInterface $tmpfs)
     {
         $uniqid = (string) preg_replace('/[^0-9a-zA-Z_]/', '', (string) $request->input('uniqid'));
+        $file = $tmpfs->readStream($uniqid);
 
-        $streamedResponse->setCallback(function () use ($tmpfs, $uniqid) {
+        $streamedResponse->setCallback(function () use ($file, $tmpfs, $uniqid) {
             // @codeCoverageIgnoreStart
             set_time_limit(0);
-            $file = $tmpfs->readStream($uniqid);
             if ($file['stream']) {
                 while (! feof($file['stream'])) {
                     echo fread($file['stream'], 1024 * 8);
@@ -168,7 +175,12 @@ class DownloadController
             'Content-Transfer-Encoding',
             'binary'
         );
-
+        if (isset($file['filesize'])) {
+            $streamedResponse->headers->set(
+                'Content-Length',
+                $file['filesize']
+            );
+        }
         // close session so we can continue streaming, note: dev is single-threaded
         $this->session->save();