+++ backend/Services/Security/Security.php
@@ -13,6 +13,7 @@ namespace Filegator\Services\Security;
 use Filegator\Kernel\Request;
 use Filegator\Kernel\Response;
 use Filegator\Services\Service;
+use Filegator\Services\Logger\LoggerInterface;
 use Symfony\Component\Security\Csrf\CsrfToken;
 use Symfony\Component\Security\Csrf\CsrfTokenManager;
 
@@ -25,25 +26,32 @@ class Security implements Service
 
     protected $response;
 
-    public function __construct(Request $request, Response $response)
+    protected $logger;
+
+    public function __construct(Request $request, Response $response, LoggerInterface $logger)
     {
         $this->request = $request;
         $this->response = $response;
+        $this->logger = $logger;
     }
 
     public function init(array $config = [])
     {
         if ($config['csrf_protection']) {
+
+            $key = isset($config['csrf_key']) ? $config['csrf_key'] : 'protection';
+
             $http_method = $this->request->getMethod();
             $csrfManager = new CsrfTokenManager();
 
             if (in_array($http_method, ['GET', 'HEAD', 'OPTIONS'])) {
-                $this->response->headers->set('X-CSRF-Token', $csrfManager->getToken('protection'));
+                $this->response->headers->set('X-CSRF-Token', $csrfManager->getToken($key));
             } else {
-                $token = new CsrfToken('protection', $this->request->headers->get('X-CSRF-Token'));
+                $token = new CsrfToken($key, $this->request->headers->get('X-CSRF-Token'));
 
                 if (! $csrfManager->isTokenValid($token)) {
-                    throw new \Exception('Csrf token not valid');
+                    $this->logger->log("Csrf token not valid");
+                    die;
                 }
             }
         }
@@ -60,6 +68,7 @@ class Security implements Service
             if (! $pass) {
                 $this->response->setStatusCode(403);
                 $this->response->send();
+                $this->logger->log("Forbidden - IP not found in allowlist ".$this->request->getClientIp());
                 die;
             }
         }
@@ -76,6 +85,7 @@ class Security implements Service
             if (! $pass) {
                 $this->response->setStatusCode(403);
                 $this->response->send();
+                $this->logger->log("Forbidden - IP matched against denylist ".$this->request->getClientIp());
                 die;
             }
         }